生成世界国家地区信息

+++ microservices-data/src/test/java/com/hkbea/data/xml/CountryRegion.java
@@ -9,7 +9,7 @@ import java.util.Set;
  */
 public class CountryRegion {
 
-	private Long id;
+	private int id;
 	
 	private String name;
 	
@@ -23,18 +23,18 @@ public class CountryRegion {
 		this.code = code;
 	}
 
-	public CountryRegion(Long id, String name, String code) {
+	public CountryRegion(int id, String name, String code) {
 		super();
 		this.id = id;
 		this.name = name;
 		this.code = code;
 	}
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
